Enhance documentation and features across various modules
- Added note on `binlog_format=ROW` validation in MySQL CDC documentation to prevent data loss.
- Updated CDC documentation to include binlog format validation and DLQ details for better error handling.
- Clarified Cedar authorization status in the Schema Registry documentation.
- Improved Zstd decompression algorithm description in the compression documentation.
- Introduced resilience features for connectors, including retry logic and DLQ configuration.
- Enhanced consumer group documentation with details on sticky partition assignment and coordination protocol.
- Updated dashboard error handling notes for clarity on CORS and replication factors.
- Expanded exactly-once semantics documentation with error classification and per-batch timeout details.
- Corrected Python examples in getting started guide for producer usage.
- Added warnings for unimplemented features in Iceberg sink documentation.
- Introduced leader election details for Rivven Operator in Kubernetes documentation.
- Enhanced connection pooling notes in RDBC documentation.
- Updated schema registry documentation to clarify authentication methods and Cedar integration.
- Added JSON Schema compatibility rules to schema registry documentation.
- Improved security documentation with TLS connection examples and enhanced security features.
- Added segment data format details in tiered storage documentation.
- Updated Python client README for producer and consumer methods with batching details.
- Enhanced type hints in Python type stubs for better clarity on producer and consumer methods.
- Improved integration test script to clean up stale Docker containers before running tests.

+++ crates/rivven-integration-tests/tests/connect_e2e.rs
@@ -691,7 +691,7 @@ async fn test_connector_batching() -> Result<()> {
         if messages.is_empty() {
             break;
         }
-        consumed += messages.len();
+        consumed += messages.len() as u32;
         offset = messages.last().unwrap().offset + 1;
     }
